Move core methods into lib
