+++ src/lib/lib.rs
@@ -1,0 +1,7 @@
+pub mod build;
+mod client;
+pub mod lock_file;
+pub mod schema_file;
+pub mod utils;
+
+pub use client::Client;
