+++ src/lib/utils/mod.rs
@@ -1,4 +1,4 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-pub mod terminal;
-pub mod print;
+pub mod files;
+pub mod key_pair;
